feat(Shuffle.svelte): add functionality to shuffle recipes when the button is clicked
The previous version of the Shuffle component only had a button with the text "Shuffle". This commit adds functionality to the component by introducing a new script block. The script block exports a function `shuffleRecipes` which will be called when the button is clicked. The button now has an `on:click` event listener that triggers the `shuffleRecipes` function. This allows users to shuffle recipes by clicking the button.