feat(recipes/[slug]/+page.js): add functionality to fetch and include recipe instructions in the page data
The changes in this commit add a new function `handleFetchInstructions` that fetches the analyzed instructions for a recipe from the Spoonacular API. The function takes a `recipeId` parameter and constructs the API URL using the provided `PUBLIC_SPOONACULAR_API_KEY`. It then makes a GET request to the API and returns the fetched data.

In the `load` function, the `handleFetchInstructions` function is called with the `recipeId` to fetch the instructions for the recipe. The fetched instructions are included in the `props` object that is returned from the `load` function.

This change improves the functionality of the page by providing the recipe instructions along with other recipe information.