feat(search/[slug]/+page.js): add new route file to handle fetching recipe data from Spoonacular API
The new file `search/[slug]/+page.js` is added to handle the fetching of recipe data from the Spoonacular API. The `load` function is implemented to fetch the recipe data based on the `params.slug` value. The function `handleFetchRecipe` is defined to construct the API URL and make the GET request to fetch the data. The fetched data is then returned as props to be used in the component.