feat(signin): redirect to home page after successful login to improve user experience
In the signin page component, after successfully receiving a token from the server, the code has been updated to redirect the user to the home page using the `goto` function from the `$app/navigation` module. This change improves the user experience by automatically navigating the user to the home page after logging in.